Refactor project structure and adapt daily_question.py for GitHub Actions
