+++ src/utils/__init__.py
@@ -1,0 +1,4 @@
+# src/utils/__init__.py
+"""Utility functions for Daily Interview Question."""
+from .question_generator import generate_behavioral_question
+from .email_sender import send_email
